requirements.txt
cassandra-driver>=3.14.0,<3.15.0
aiohttp>=3.3.1,<3.4.0